add sei-chein info and refactoring
