+++ chains/mainnet/dymention.json
@@ -1,8 +1,8 @@
 {
     "chain_name": "dymension",
     "coingecko": "dymension",
-    "api":["https://us-central1-omakase-explorer.cloudfunctions.net/proxyToSentryApi"],
-    "rpc": ["https://us-central1-omakase-explorer.cloudfunctions.net/proxyToSentryRpc"],
+    "api":["https://us-central1-omakase-explorer.cloudfunctions.net/proxyToSentryDymApi"],
+    "rpc": ["https://us-central1-omakase-explorer.cloudfunctions.net/proxyToSentryDymRpc"],
     "snapshot_provider": "",
     "coin_type": 60,
     "sdk_version": "0.46.10",
